requirements.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-audiofield==0.10.0
django-bootstrap4==1.1.1
django-celery==3.3.1
django-cors-headers==3.4.0
django-geoposition==0.3.0
django-google-maps==0.12.1
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.10
kiwisolver==1.2.0
kombu==4.6.11
lxml==4.5.2
matplotlib==3.3.1
MouseInfo==0.1.3
numpy==1.19.0
opencv-python==4.3.0.36
pandas==1.1.0
Pillow==7.2.0
psycopg2==2.8.4
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
response==0.4.0
scipy==1.5.2
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.30
whitenoise==5.0.1
